Use number of files printed to screen as page size
Number of lines is a good approximation when sorting files by names, when (usually) there are few directory transitions.
Other sort orders might result in a directory transition every file, so page down skips too many files.
This is still not perfect, since the number of transition in the displayed page might be different than in the new page, but it is better.